Added Health Bar and Win/Lose Logic
Added a simple health bar to the enemy.

I also added the logic for the player to win or lose. Currently, the enemy cannot do dmg, so the player can never lose, but the logic is there for that too. Everything just results in a debug log for this at the moment.