Fixed the jump and add ik constraints for knees
The enemy/dummy is now rigged to react to hits. Add Ik_Maxs to allow the maximum displacement to be set for ik targets. This is automatically checked by the 'Hand' script.

Adding colliders to the feet and hands broke the jumping for the character, but everything is now back to normal.