better ML-agent
Rebuilt the enemy controller from scratch (i.e. not inheriting from ThirdPersonController). This resulted in a much better agent behavior, but there is still a flaw in how the player attacks. The agent managed to find a bug where mearly standing really close to the target essentially shoves the right hand into the enemy triggering the collision, and therefore the punch logic.